+++ Sources/SwiftyHaru/Graphics/PDFColorSpace.swift
@@ -11,7 +11,7 @@ import CLibHaru
 #endif
 
 /// A profile that specifies how to interpret a color value for display.
-public enum PDFColorSpace: UInt32 {
+public enum PDFColorSpace: UInt32, CaseIterable {
     case deviceGray
     case deviceRGB
     case deviceCMYK